Shim SlippiDirectCodes to Rust port.
This changes out the internals of the `SlippiDirectCodes` class, wherein
it'll now just silently call over to the Rust side. This keeps the
changes minimal for now as we continue to migrate things out.

This notably lacks the change that Ishiiruka had to determine the user
config folder path, since it appears the mainline build doesn't need it?
Might need further testing there.

This compiles but has not been fully tested/vetted yet. Building this
branch currently requires the corresponding branch from
slippi-rust-extensions.